backend/requirements.txt
opencv-python-headless
numpy
fastapi
python-dotenv
httpx
pillow
pydantic
uvicorn[standard]
